+++ source/georef/street_network.cpp
@@ -43,7 +43,7 @@ namespace navitia { namespace georef {
 const auto source_e = ProjectionData::Direction::Source;
 const auto target_e = ProjectionData::Direction::Target;
 
-navitia::time_duration PathFinder::crow_fly_duration(const double distance) const {
+navitia::time_duration PathFinder::crow_fly_duration(const float distance) const {
     // For BSS we want the default speed of walking, because on extremities we walk !
     const auto mode_ = mode == nt::Mode_e::Bss ? nt::Mode_e::Walking : mode;
     return navitia::seconds(distance / (default_speed[mode_] * speed_factor));
@@ -312,7 +312,7 @@ PathFinder::crow_fly_find_nearest_stop_points(const navitia::time_duration& max_
                                               const proximitylist::ProximityList<type::idx_t>& pl) {
     // Searching for all the elements that are less than radius meters awyway in crow fly
     float crow_fly_dist = max_duration.total_seconds() * speed_factor * georef::default_speed[mode];
-    return pl.find_within(start_coord, crow_fly_dist);
+    return pl.find_within(start_coord, double(crow_fly_dist));
 }
 
 
@@ -770,7 +770,7 @@ static edge_t get_best_edge(vertex_t u, vertex_t v, const GeoRef& georef) {
 Path create_path(const GeoRef& geo_ref,
                  const std::vector<vertex_t>& reverse_path,
                  bool add_one_elt,
-                 double speed_factor) {
+                 float speed_factor) {
     Path p;
 
     // On reparcourt tout dans le bon ordre
